+++ packages/api/internal/handlers/sandbox.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package handlers
 
 import (
