+++ packages/orchestrator/internal/sandbox/nbd/path_direct.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package nbd
 
 import (
