+++ packages/orchestrator/internal/sandbox/network/network.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package network
 
 import (
