--- packages/orchestrator/internal/sandbox/network/network_other.go
@@ -1,16 +1,0 @@
-//go:build !linux
-// +build !linux
-
-package network
-
-import (
-	"errors"
-)
-
-func (s *Slot) CreateNetwork() error {
-	return errors.New("platform does not support network creation")
-}
-
-func (s *Slot) RemoveNetwork() error {
-	return errors.New("platform does not support network removal")
-}
