+++ packages/orchestrator/internal/sandbox/sandbox.go
@@ -1,6 +1,3 @@
-//go:build linux
-// +build linux
-
 package sandbox
 
 import (
