--- packages/orchestrator/internal/sandbox/sandbox_other.go
@@ -1,99 +1,0 @@
-//go:build !linux
-// +build !linux
-
-package sandbox
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"net/http"
-	"time"
-
-	"go.opentelemetry.io/otel/trace"
-
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/dns"
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/build"
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/network"
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/template"
-	"github.com/e2b-dev/infra/packages/shared/pkg/chdb"
-	"github.com/e2b-dev/infra/packages/shared/pkg/grpc/orchestrator"
-	sbxlogger "github.com/e2b-dev/infra/packages/shared/pkg/logger/sandbox"
-	"github.com/e2b-dev/infra/packages/shared/pkg/storage"
-	"github.com/e2b-dev/infra/packages/shared/pkg/storage/header"
-)
-
-var httpClient = http.Client{
-	Timeout: 10 * time.Second,
-}
-
-type Sandbox struct {
-	Config *orchestrator.SandboxConfig
-}
-
-func (s *Sandbox) LoggerMetadata() sbxlogger.SandboxMetadata {
-	return sbxlogger.SandboxMetadata{
-		SandboxID:  s.Config.SandboxId,
-		TemplateID: s.Config.TemplateId,
-		TeamID:     s.Config.TeamId,
-	}
-}
-
-// Run cleanup functions for the already initialized resources if there is any error or after you are done with the started sandbox.
-func NewSandbox(
-	ctx context.Context,
-	tracer trace.Tracer,
-	dns *dns.DNS,
-	networkPool *network.Pool,
-	templateCache *template.Cache,
-	config *orchestrator.SandboxConfig,
-	traceID string,
-	startedAt time.Time,
-	endAt time.Time,
-	isSnapshot bool,
-	baseTemplateID string,
-	clickhouseStore chdb.Store,
-) (*Sandbox, *Cleanup, error) {
-	return nil, nil, errors.New("platform does not support sandbox")
-}
-
-func (s *Sandbox) Wait() error {
-	select {
-	case fcErr := <-s.process.Exit:
-		stopErr := s.Stop()
-		uffdErr := <-s.uffdExit
-
-		return errors.Join(fcErr, stopErr, uffdErr)
-	case uffdErr := <-s.uffdExit:
-		stopErr := s.Stop()
-		fcErr := <-s.process.Exit
-
-		return errors.Join(uffdErr, stopErr, fcErr)
-	}
-}
-
-func (s *Sandbox) Stop() error {
-	err := s.cleanup.Run()
-	if err != nil {
-		return fmt.Errorf("failed to stop sandbox: %w", err)
-	}
-
-	return nil
-}
-
-func (s *Sandbox) Snapshot(
-	ctx context.Context,
-	tracer trace.Tracer,
-	snapshotTemplateFiles *storage.TemplateCacheFiles,
-	releaseLock func(),
-) (*Snapshot, error) {
-	return nil, errors.New("platform does not support snapshot")
-}
-
-type Snapshot struct {
-	MemfileDiff       build.Diff
-	MemfileDiffHeader *header.Header
-	RootfsDiff        build.Diff
-	RootfsDiffHeader  *header.Header
-	Snapfile          *template.LocalFile
-}
