--- packages/orchestrator/internal/sandbox/uffd/constants.go
@@ -1,5 +1,0 @@
-//go:build linux
-// +build linux
-
-package uffd
-
