--- packages/orchestrator/internal/sandbox/uffd/serve_other.go
@@ -1,36 +1,0 @@
-//go:build !linux
-// +build !linux
-
-package uffd
-
-import (
-	"errors"
-	"fmt"
-
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/block"
-)
-
-var ErrUnexpectedEventType = errors.New("unexpected event type")
-
-type GuestRegionUffdMapping struct {
-	BaseHostVirtAddr uintptr `json:"base_host_virt_addr"`
-	Size             uintptr `json:"size"`
-	Offset           uintptr `json:"offset"`
-	PageSize         uintptr `json:"page_size_kib"`
-}
-
-func getMapping(addr uintptr, mappings []GuestRegionUffdMapping) (*GuestRegionUffdMapping, error) {
-	for _, m := range mappings {
-		if !(addr >= m.BaseHostVirtAddr && addr < m.BaseHostVirtAddr+m.Size) {
-			continue
-		}
-
-		return &m, nil
-	}
-
-	return nil, fmt.Errorf("address %d not found in any mapping", addr)
-}
-
-func Serve(uffd int, mappings []GuestRegionUffdMapping, src *block.TrackedSliceDevice, fd uintptr, stop func() error, sandboxId string) error {
-	return errors.New("platform does not support UFFD")
-}
